@@ -21,9 +33,9 @@
     "prop-types": "^15.7.2",
     "query-string": "^6.13.1",
     "rc-color-picker": "^1.2.6",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-headroom": "^3.0.0",
     "react-image": "^4.0.3",
     "react-image-crop": "^8.6.12",
